Create LICENSE.md
This commit adds the LICENCE.md file to the repository. The license file contains the full text of the software license under which the project is released. It includes information about permissions, restrictions, and obligations for using, modifying, and distributing the code. By including a license file, we ensure that the project is properly licensed and can be shared and used by others in compliance with the chosen license.